Creating common public handlers
Only for a few services  - if we like this pattern, we'll do more.
